+++ plugins/partiql-memory/src/main/kotlin/org/partiql/plugins/memory/MemoryPlugin.kt
@@ -7,8 +7,8 @@ import org.partiql.spi.function.PartiQLFunctionExperimental
 
 class MemoryPlugin(private val catalogs: Map<String, MemoryConnector>) : Plugin {
 
-    override fun getConnectorFactories(): List<Connector.Factory> = listOf(MemoryConnector.Factory(catalogs))
+    override val factory: Connector.Factory = MemoryConnector.Factory(catalogs)
 
-    @PartiQLFunctionExperimental
-    override fun getFunctions(): List<PartiQLFunction> = emptyList()
+    @OptIn(PartiQLFunctionExperimental::class)
+    override val functions: List<PartiQLFunction> = emptyList()
 }